update the generic bound logic in the derive macro
This rewrites the type bound logic after having analyzed it again and
after the changes due to adding the cont generic in `Miniconf`.

The Miniconf<Y> const generic Y is now the miniconf recursion depth
again. And it's an upper bound to the state size.
Before, the idea of making it exactly the key length war broken:
We had `impl Miniconf<0> for Option<T> where T: SerDe`
but but `#[miniconf(defer(1)) a: [Option<T>; N]` it would not use the
miniconf API of `Option<T>` as desired but it would for
`#[miniconf(defer(0)) a: Option<T>`.
Using the recursion depth as indicator which API to call fixes this.

At the cost of the state size now being potentially a bit too large.